- Updated Recall Partner to have 20secs in the ground before recalling partner, also skill is blocked from being recasted during that time. - Added target "noone" to skill_unit_db (used only by the Wedding recall skill) - Cleaned up Recall Guild to always work in castles, changed the gmaster_flag from int to a struct guild pointer.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3540 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/guild.c
@@ -499,7 +499,7 @@ int guild_recv_info(struct guild *sg)
 				if (guild_checkskill(sg, skill_num[i]))
 					pc_blockskill_start(sd, skill_num[i], 300000);
 			//Also set the guild master flag.
-			sd->state.gmaster_flag = (int) g;
+			sd->state.gmaster_flag = g;
 		}
 	}else
 		before=*g;
@@ -1501,7 +1501,7 @@ int guild_gm_changed(int guild_id, int pos)
 	if (g->member[0].sd && g->member[0].sd->fd)
 	{
 		clif_displaymessage(g->member[0].sd->fd, "You have become the Guild Master!");
-		g->member[0].sd->state.gmaster_flag = (int) g;
+		g->member[0].sd->state.gmaster_flag = g;
 	}	
 	return 1;
 }
